AUTH-1405: use String for uri rather than Java URI
Seeing errors when trying to parse strings to URI on Windows.  Given that the URI are consumed as String by Selenium anyway we can just remove the conversion.  There's some loss in not being able to retrieve component parts of the URI for testing, but this lets us move forward with running on Windows.