BAU: Remove dependencies on Account Management
As AM has moved to a different team we should no longer own the AM tests nor depend on AM as a client, as external changes to AM can break the acceptance tests without Auth knowing.

Some scenarios have been deleted temporarily as they depend on users being deleted in AM to run.  These will be restored when additional test users have been created.

There are two remaining dependencies to be removed which only use AM to redirect to sign in, but will not break at the current time.