AUT-1320: Use the single reset test data script in the pipeline
Create a further separate script with a function for resetting test users only.
Allows the script to be called by the pipelines which get their AWS credentials and environment variables in a different way to the local config.
Call the script from the pipeline so that now the single test data creation and reset script can be used both in the pipeline and locally.