Decouple shell scripts from gds-cli
This change removes the dependency on the gds-cli tool from the
run-acceptance-tests.sh and reset-test-data.sh scripts.

Ideally, `run-acceptance-tests.sh` should be run on the CLI with
`gds aws xxx -- ./run-acceptance-tests.sh`, rather than calling out
to `gds` from within the script. This allows the script to be run with
credentials directly from environment variables (ie. sourced from the
'Command line or programmatic access' section of the AWS SSO portal).

`run-tests-locally-gds-cli.sh` is a helper script that can be used to
run `gds aws digital-identity-dev -- ./run-acceptance-tests.sh -l` with
a single command.