BAU: Clear cookies to simulate logout
Some tests are getting stuck waiting for logout to complete on the RP stub in the pipeline, which is causing tests to fail.  A workaround was to stop waiting for logout to complete as this was the last step in most tests.  However, where there was a dependency on logout completing for the next scenario to continue this could not be done.

Instead do not wait for logout to complete before the next step, but clear cookies so that the user's session is effectively cleared, and they will start the journey again.