BAU: Allow initial RP stub step to retry
As we're not actually testing the RP stub in these acceptance tests,
we are able to retry a limited number of times if session creation
fails.

The retry limit is global across all features, and will allow 5 failed
session initiation attempts.

Hopefully this will allow us to prevent some of the occasional stub
errors from causing a complete acceptance test failure.

There is a limit in place for retries to ensure that we do eventually
fail out if there actually is an issue, rather than potentially running
until the end of time.

This limit is global to ensure that actual flaky behaviour is not
always ignored.

This is potentially a risk, because we may end up discarding certain
valid flaky issues in the OIDC flow, but in theory, the smoke tests will
catch this behaviour still.