BAU: perform test-data initiation in parallel
Rather than sequentially initiating the test data, we can run certain
tasks at the same time.

`&` has been added to the end of most test data function calls, grouped
with `wait` statements.

This allows tasks that can be run simultaneously to be run as such,
while ensuring that tasks that perform future dynamodb row modification
are still run after the initialisation step.

This dramatically reduces the overall run time of the tests.

This should be replaced with on-the-fly test user creation in the
future.