BAU: Don't delete cookies when retrying the stub
Cookies were being deleted before out of an abundance of caution.

This is actually disruptive in the case of SSO uplift - we need
to keep the existing session in place.

In theory, if we are retrying the RP stub, it means that no
session stuff has actually happened at that point, so there's no
actual need to clear them when retrying.

This has the side effect of fixing the uplift edge case without
writing any more java.