AUT-3800: Create acceptance test users on the fly
Previously, we created the users at once before running the tests.
This prevented tests from running simultaneously, as the email addresses
were not unique per-run.

User data is now created for each test, and the email addresses are
generated using the timestamp of test-startup, and a counter, which
increments after each user is created.

Test configuration is now loaded from AWS SSM, and the values are set
when authentication-api is deployed.