AUT-4072: Add ability to run API tests
We need to write acceptance tests against the Account Management API
starting with the new MFA Methods Management endpoint.  This repo has
some hooks defined that are very specific for testing though the UI so
the tests have been updated to either have a tag of @UI or @API so that
the API tests can be excluded from the hooks.

A new feature file has been added for the mfa-methods endpoint with a
very simple call to the API to prove how to call API endpoints.

The `openapi-specs` repo allows Imposter to be run to provide the
example responses defined in the OpenAPI spec as used to prove the first
acceptance test locally.  We need a mock server for the mfa-methods
endpoint as this API is a private API and not easily accessed from a
developers' laptop.