BAU: Fix local running of tests against old dev environment
The acceptance tests weren't running correctly against the dev environment
due to the .env file not being created from SSM.

Added a step into the Docker file to generate a .env file and copy it into
the docker image when running against the old dev environment.

Dropped the code that created the new env file with a timestamp to make
it easier to copy the generated file.