ATO-1644: Tweak multitab reauth tests for no shared session
The previous multi-tab re-auth test implicitly relied on the shared
session. This was because once authenticated in tab 1, opening a new
tab and doing a silent login in tab 2 would overwrite the gs cookie
with a new session ID. This over-written gs cookie applies to both
tabs, and therefore when the user logs out in tab 1 this destroys
their orchestration session and their shared redis session associated
with the new session id from tab 2. Then trying to continue their
reauth journey results in an error due the the fact that they no
longer have a shared redis session available.

When auth no longer reads the shared redis session, this no-session
response does not happen. There is still an auth session available as
logouts no do not destroy the auth session. This means a user is able
to continue their re-auth journey up until the point they are handed
back to orchestration. The error page is now rendered here as they do
not have an orchestration session and are therefore redirected to
the "There’s a problem with this service" page.