BAU: Standardise use of SSM
The acceptance tests have the ability to check SSM if
a required configuration value is not present in an
environment variable.  This only needed when running the
tests from an IDE as all other ways, on the pipeline
or locally via docker, expect an .env file to be
present.

Each of the run scripts will now create an .env file
from SSM if one has not been provided to the docker
image as a shared volume.  This allows the local run
to create a bespoke .env file with local overrides
such as a more limited set of cucumber filter tags
for example.