BAU: minor improvements to rundocker.sh
- Instead of partially hardcoding the selenium base image, instead allow
  the user to specify the whole image name + version as an argument.
  This ensures that if no image is specified, the default *from the
  Dockerfile* is used. As-was, we would have to remember to bump the
  version both in this script and the Dockerfile, which is error-prone.
- Remove `--privileged` flag from `docker run` commands, as it is not
  needed for these tests and can pose security risks.

NB: weird-looking `"${BUILD_ARG_ARGS[@]+"${BUILD_ARG_ARGS[@]}"}"` syntax
is required for bash 3.2 (macOS default) compatibility. With that bash
version, if this arcane syntax is not used and the array is empty, it
will not expand to anything, which causes the script to exit (because
`set -u` is used) with `BUILD_ARG_ARGS[@]: unbound variable` error.