AUT-4863: Add counter to email assessment email addresses
Increase the likelihood of collisions by adding the counter to the high risk email addresses generated for tests. This is important to ensure that when the tests are run they are not served a cached risk assessment decision from a reused email address. It also ensures we continue the established pattern of using counters as well as timestamps in acceptance test emails.