AUT-4863: Add ability to wait some seconds
The email blocks feature takes a little time to perform the email check against an email address from the moment it is submitted to when the the decision is checked.

Therefore we should wait a little in order to guarantee a DENY decision is made when we expect to be blocked.