AUT-4863: Assert email block works with `/update-email` API
Using `RestAssured`, create a step to make a request to the `/update-email` API and steps to check different responses

This allows us to assert that emails are blocked or allowed as expected.

Doing this, rather than directly invoking the lambdas, means we need to provide the hostname for the account management APIs. This is requested from a new config variable `AM_API_URL`.

As noted in the README, running the API tests locally require setting the `AM_API_URL` as an override and using the AM API proxy to get tunnel though VPC restrictions.

The URL will also need setting up in each environments parameter store.