AUT-4574: Correct broken cross-tab logout reauth tests
We should be asserting that when we log out that we can no longer do reauth in the other tab. Our definition of logout here is "delete all cookies". This is fine, however the tests do not reflect what error we would show and when. Instead of a problem with the service, we show the page has expired page when the session has gone.

Tests have also been corrected to expect the error at the right place, as the test scenarios state "when they enter a valid email" and "when they enter a valid password"